Avoid compiler warning.

+++ OpenSim/Tools/MarkerPlacer.cpp
@@ -264,7 +264,7 @@ bool MarkerPlacer::processModel(Model* aModel,
 
     const auto avgRow = staticPoseTable.averageRow(_timeRange[0],
                                                    _timeRange[1]);
-    for(int r = staticPoseTable.getNumRows() - 1; r > 0; --r)
+    for(size_t r = staticPoseTable.getNumRows() - 1; r > 0; --r)
         staticPoseTable.removeRowAtIndex(r);
     staticPoseTable.updRowAtIndex(0) = avgRow;
